- Remove warpaffine related code and replace it with letterbox for image
scaling and padding
- Refactor the Transform structure to simplify the logic for calculating
transformation matrices
- Optimize the mask processing logic by removing the edge cropping step
- Update the relevant interface calls in the backend to adapt to the new
letterbox implementation

+++ modules/trtyolo/infer/trtyolo.hpp
@@ -67,9 +67,9 @@ struct TRTYOLOAPI Image {
  * @brief 掩码结构体，用于存储掩码数据及其尺寸信息
  */
 struct TRTYOLOAPI Mask {
-    std::vector<uint8_t> data;        // < 掩码数据
-    int                  width  = 0;  // < 掩码宽度
-    int                  height = 0;  // < 掩码高度
+    std::vector<float> data;        // < 掩码数据
+    int                width  = 0;  // < 掩码宽度
+    int                height = 0;  // < 掩码高度
 
     /**
      * @brief 构造函数，初始化掩码尺寸并分配数据空间
